+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 #### Miscellaneous
 * Added exception `SelectNotFound`
 * The `Page` class has been added to `__init__` and is now available for use
+* If the `method` parameter in either the `ReactionMenu` or `ViewMenu` constructor is of the wrong type, `IncorrectType` is now raised